adjusting the printout messages.
I think the single, simultaneous pass approach is superior so I am making that the default and renaming the old method relax_to_barotrope_old, for now, probably to be removed at some point in the future. The difference, besides the adjusted printouts, is that the single pass looks for the max of dJ and var(drho) for the tolerance comparison.