Fixed bug introduced in commit 60490ec58ab3e0072705396fff672792e6008334
When I let cmp.relax_to_barotrope skip the cms.update_zetas part if
qrot==0 I forgot that this also skips setting the flag to recalculate
the potential, for the purpose of recalculating the pressure, for the
purpose of recalculate the density. So cmp.update_densities was not
computed correctly. The fix was to call cms.update_Js, even though we
know all Js are 0 in this case. The call resets the right flags and the
pressure/density relation can be computed correctly, and very quickly
for a non-rotating planet.