Fixed bug introduced in commit 220a7ed52df0057572038095e3732623ac3c7a15 [formerly 60490ec58ab3e0072705396fff672792e6008334]
When I let cmp.relax_to_barotrope skip the cms.update_zetas part if
qrot==0 I forgot that this also skips setting the flag to recalculate
the potential, for the purpose of recalculating the pressure, for the
purpose of recalculate the density. So cmp.update_densities was not
computed correctly. The fix was to call cms.update_Js, even though we
know all Js are 0 in this case. The call resets the right flags and the
pressure/density relation can be computed correctly, and very quickly
for a non-rotating planet.


Former-commit-id: 5f1dbe2ed4b72324727b26c78dbb78c80bb7df97