Fix to transition radii in +models (#25)
This commit changes the placement of transition layers in models from the `+models` package.

The affected models are `models.double_polytrope`, `models.double_polytrope_w_core`, and `models.triple_polytrope`. The transition normalized radii are handled in a (very subtly) different way. The transition happens at the layer with radius nearest to requested value rather than the first layer below it. And the next eos is applied _at_ the transition radius instead of the next layer.

In addition an optional parameter `forcematch` forces one of the spheroid lambdas to the exct value of the requested transition. This works better for models with a small number of layers. At the moment forcematch=true must be specified as the 4th argument to relevant models but in the future we will probably make this behavior the default. 

We also added some more input validation and better error messages to all models.