New lambda spacing option for all models in +models:
Models now accept also a function handle for the lamstrat argument.
This function should take a single scalar integre (N) and return a
vector of length N appropriate for normalized layer radii. The output
will be sorted if it isn't already. For models with a constant density
core the bottom of the lambda spacing should not be too far from the
requested core radius of the model. (This will be verified and partly
enforced.)

This is the prefered way to specify lambda spacing, allowing full
control. The old way is preserved for compatability.

Examples:
1) Request equally spaced layers throughout the planet:
    f = @(N)linspace(1, 1/N, N);
    cmp = models.single_polytrope(32, [1e5, 1], f);

2) Request equally spaced layers in throughout the envelope:
   f = @(N)linspace(1, rcore/R, N);
   cmp = models.single_polytrope_w_core(32, [1e5, 1, 9000, 0.3], f);

NOTES:
1) When using the lamstrat fucntion handle the resulting lambda spacing
is not modified by the model construction except perhaps to force match
transition radii to be on layer radii (if forcematch==true). If we want
to keep the half-width surface behavior we have to expressly put it in
the lamstrat function.

2) The interpretation of the old-style lamstrat 2-vector changed somewhat
for the *_w_core models. The vector lamstrat=[x,y] now reads as putting a
fraction x of layers in the top fraction y of the ENVELOPE, not the planet.


Former-commit-id: 1763269848c2d585a961e877075d549b9581b93b