Starting set of lambda distributions package.
We put a set of more-or-less standardized functions that make
lambda arrays that can be assigned to cms.lambdas, using different
strategies. This commit just starts the set so that we can establish
the standard and usage but none of the available strategies is really
very good. Following commits will include better ones. Eventually we
expect to settle on one or two "best" lambda strategies.