Minor overhaul of +lambdas
Cleaned up a bit based on tests performed in the lamstrat_new branch.
The mysterious half-width layer is only relevant to lambdas with
otherwise constant spacing near the surface, so for now only in
equal_dr.m and topheavy.m where it is in fact crucially important for
reasons still unknown. It is better calculated now in those lambdas
and removed from the rest.

The recent tests also show how most of the methods in +lambdas are
not useful (this was expeted) and they will soon be removed.


Former-commit-id: c16a5dfe43649a6bf530777424ccdb77e60ffa09