Another small cleanup of +lambdas package.
Moved not helpful distributions to a +unused subpackage, in
preparation for removal.

Added a trizone distribution.

best.m points to topheavy.m with [3/4, 1/2] skew, for now.

The saga of lambda strategies is far from over but we are getting
near a point where we can leave it alone for a while and work with
what we have.


Former-commit-id: 8bf00083fd31066352e1b27ac07684ccbffc60e5