End (for now) work on lambda distributions.
This is a good stopping point. So far eveything we
try is worse than the top-heavy [3/4, 1/2] distribution.
Making this the default and moving on until beter ideas
emerge.

Keeping the lamstrat_new branch around for now just as
a place to store the tests in case we care to look at them


Former-commit-id: 8e7fdd02f8cbdf043a98b0af729cf331f066b153