No-barotrope ("pure") models in +models.
The pure_cms and pure_cmp model-building functions are
of course not needed for convenience. All they do is
assign the lambdas/deltas or ai/rhoi fields. But the point
is that some scripts (notably the drivers in CMSPIPE) expect
a handle to a model-building function. So we supply one.

In truth the pure_cmp function is not entirely redundant.
It's not the easiest way to initialize a cmp given radius and
density vectors, since it calculates the total mass as well.


Former-commit-id: f44efdba6a5566092f0d5b8852c215660ccb4660