Fixed bug in nonrotating density-only calculation.
The fast density-only update branch of CMP.relax_to_barotrope
used to skip both cms.update_zetas and cms.update_Js. However
at least a partial update to Js is required. A spherical planet
still has non-zero internal J values (in the tilde and primed
tilde variants). This commit restores a call to cms.update_Js
in the q=0 branch of cmp.relax_to_barotrope. This significantly
slows down the calculation but results in correct density structure.

In the future a partial update to Js for a non rotating planet is
possible. Basically a copy of the J allocation as in InitCMS, but
with a density input.


Former-commit-id: 73f987cb325dc0035ff95a3ff06e4fc912e84bf3