Fixed calculation of Js for spherical planet.
A bug in the private CMS.allocate_spherical_Js miscalculated
the double primed Js (eq. 48d in CMS.pdf or eq.27 in Hubbard 2013)
for the case of a spherical body. It's the messy mixe up of tilded
and non-tilded variables, the unfortunate legacy of starting this
project before I fully rewrote the theory in clean notation.

 * A new public method CMS.update_spherical_Js follows the layout
   of the usual update_Js but calls the private allocate_spherical_Js
   to do the update. The Js actually change enough to require almost
   as many iterations as for a rotating planet, but each iteration is
   much faster.

 * The density-only branch of CMP.relax_to_barotrope has been updated
   to use this new method.


Former-commit-id: bb8087172e18ee5d008b6c9afe21c30dcb5bffbc