Copyable classes.
Made ConcentricMaclaurinSpheroids, CMSPlaner, and Barotrope derive
from matlab.mixin.Copyable imstead of plain handle. The thus
inherit the copy method that performs a shallow copy with all the
desired features (e.g. no dependables, no calling constructors or
set methods etc.)

In CMSPlanet also implement an override of the copyElement method.
This override calls the copy method again on the obj.cms and
obj.eos fields, thus implementing a (selective) deep copy, which is
precisely what we wanted.

This change was rather easy considering it's a change of the
superclass of everything! Tests so far indicate everything behaves
as expected.

ONE CAVEAT: when we create a cmp using our template models we set
the eos field to be a heterogeneous array of barotrope handles.
This saves memory since we have an eos for each layer but usually
2-4 eoss total for hundreds of layers. However when we make a deep
copy the new eos field will be populate for a new copy of eos(i)
for each layer! This is usually not an issue, unless we start
using tabular eoss that take up a lot of memory. If we do then the
solution is simply to create a new cmp object and then copy just
the cms field from the old object.


Former-commit-id: 274bf286caf292cfe02e049244a3a85bb5909afa