Implemented a downsampler.
The static method CMSPlanet.updownsample(incmp,N) takes the input
cmp object (or a preconverted struct) and constructs an output
cmp object with only N layers trying to most closely resemble the
original. The layers are not interpolated! Instead the layers are
placed where an old layer used to be, and given densities such
that the mass of a new super layer is equal to the total mass of
the original layers it replaces. Shape is not copied, so the new
object needs to be re-relaxed before it will actually resemble
the old one. The qrot, opts, and eos are copied but we
usually won't rely on this and instead set the parameters
explicitly.

A placeholder for up-sampling (N greater than input nlayers)
exists but I'm not sure it will ever be needed.