New methods tof2cms and tof2cmp.
To facilitate comparison between Theory of Figures and CMS we have
the methods CMS.tof2cms(svec,rhovec,q) and
CMP.tof2cmp(svec,rhovec,q). The CMS iterates cms.relax() until a
set of lambdas is found that leads to ss/ss(1) array closely
matching the input svec/svec(1). The units of svec and rhovec are
unimportant. The CMP methods first calls the CMS method to generate
a cms, and then uses the information in svec and rhovec (now
assumed to have meaningful units) and the returned cms to calculate
cmp.M, cmp.a0, and cmp.rhoi.

Tested with a CEPAM-based model this seems to work well. In the
future we may wish to add a second level iteration to find q from
a given m.

Note that many grid-based methods use a grid point at s=0 in which
case the cms/cmp will have N-1 layers.