Squashed commit of branch pycms.
A python implementation of standalone cms calculator. The
interface is the same as the ToF calculator, almost. Basic usage is
 cms.cms(zvec,dvec,qrot,dJtol=1e-6,maxiter=100,xlayers=64,J0s=None)
Notice how the skip-n-spline parameter is specified differently from
cms.m though: instead of specifying the skip we give the desired
number of layers to solve explicitly, in xlayers. I now find this
a little more intuitive, and I may soon modify cms.m (and tof4.*)
to match.

As of this commit, the python implementation is validated against
cms.m and found to be robust and precise. But it's slow. Very, very
very, very, very slow. Basic profiling revealed nothing obvious.
The time is spent entirely in cms._eq52(), as it should. I did
try rewriting _eq52() with listcomps, instead of the more readable
for-loops, and found it did absolutely nothing. I guess this is
good news. But, as it stands, the cms.py is essentially unusable!

I am merging at this point and going to open a new branch devoted
to optimizing. First the python implementation to bring it closer
to usable, and then perhaps both implementations can benefit from
some serious algorithmic work.