parforized skipnsplie in cms.m.
This is just a first stab at reinstating parallelism. I need
something to allow a few fminsearch runs for a Saturn project. No
doubt we will return to this and do a better job soon, maybe even
move from parfor to spmd, and maybe implement a streamlined and/or
pythonized version of the fminsearch drivers (but that is project
specific). In this commit all I did was add par in front of the
for on the layer loop. I did try it on the inner, angle loop but
that was a disaster.