Implement a prerat flag to control precaclculation of lamratpow.
The CMS algorithm (eq. 52 in my notes) needs powers of ratios of
lambdas (layer radii) in a handful of lines. These equations are
called a cranjillion times overall. Since lambda values are
constant for a given model (in fact virtually constant in a
multi-poject span probably) it makes sense to precalculate these
powers. On single CPU runs this indeed leads to huge speed gain,
consistant with array lookup being faster than a single division
followed by single power-raise.

However, the lamratpow array of 32xNxN takes up some space. For N
of 1024 its about 250MB, and scales with N^2. Okay fine, but if
we want to run with a parpool it can become a problem. Data
transfer is slow enough to really hit performance or we even run
out of memory (remember each pool worker is a matlab engine!). I
tried some tricks like memoizing to amortize the data transfer or
switching to spmd and using a codistributed array or using a const
or global or whatever, but these all end up net losses. Instead,
we have a semi-documented flag to cms(...,prerat) default true to
control pre-calculation of lamratpow. It's a case by case call,
where the more cpus you use the more likely it is advantageous to
disable prerat. Even when it is a net gain it a huge disincentive
to parpool. But that's the way it is, at least for now.