Mini overhaul with minor compatability considerations.
The major improvements in this commit are:

* Change the way we control the skip-n-spline algorithm from
  specifying a number of layers to skip to specifying the number or
  the list of layers to explicitly calculate. The name of the
  parameter consequently changed from sskip to xlayers, default 64.
  A scalar xlayers means we want that many explicit layers, evenly
  distributed among the full list of density layers. Alternatively,
  a vector xlayers specifies the indices of layers to be fully
  calculated, giving us fine control. This may become critical when
  we start using the skip-n-spline much more aggressively in
  conjuction with hand-tailored layer distributions.

* Since the list of optional arguments to cms() was growing I judged
  it prudent to switch the call signature to NAME/VALUE.

Clearly, the above changes, although essentially minor/cosmetic, do
break backward compatability. The change is mostly limited to calls
to the low-level cms.m, so that users who primarily work through the
CMSPlanet class and control the settings through cmsset would be
minimaly affected. If they get an error from cmsset and read the
message without freaking out they can easily change, e.g., sskip=32
to xlayers=64 and be done. Hopefully this will not create too much
confusion.

But in case it does, this commit will be tagged so that users can
easily downgrade to an older version.