Hack for non-rotating planets in cms.m.
If cms is called with qrot=0 then the convergence test, relying on
relative difference in Js between successive iterations, will fail.
A simple hack is to squash the dJs where the calculated Js are only
numerical noise. The the algorithm can complete and the function can
return with all the expected outputs; which sometimes we want even
if we already know that Js = [-1 0 0 0 ...].