More versatile MoI calculation.
CMP.NMoI is a "pseudo access method" instead of dependent variable.
This allows us to specify if we want total MoI or spheroid MoI or
a cumsum.

Also fixed a bug in plot_spheroid_J_contribution() and added a
plot_moi_contribution().

Also pasted in gauleg() and Pn() as class-related to make
distribution to coleagues easier.