Squashed commit of optimization work:
The bottom line is that with the new xlayer scheme the previously
unimportant _update_Js was contributing and in some cases dominating
the run time. It is not optimized and overall runtime is O(nx^2) as
it should be.

Specifically:
 * The loops in _update_Js were offloaded to a @jit optimized _eqs48
 * Less important - np.dot(a,b) is faster than np.sum(a*b) and much
   faster than sum(a*b)
 * The lamratpow calculation is also offloaded to a @jitted function,
   otherwise it can dominate overall runtime for very large N
 * Making sure root_scalar uses brentq helps some. Apparently the
   loss of use of a previous value for starting guess is compensated
*  The root_scalar call was replaced with a direct call to brentq
   which helps a little, maybe due to overhead

I still fail to see convergence of J2 with increasing N to the level
promised by CMS. Eventually, with N~10^5, the J updates become
dominant again and the runtime scales with N instead of with nx*nx,
and of course becomes unacceptably long, while still not providing
convergence. But that's a story for another branch.