Squashed commit of a cleanup/perioding branch:
Main change is switching CMSPlanet to rotation
period based instead of rotation parameter. I
recently decided this is the only way to maintain
consistency in all situations. This is not as
crucial as with a TOFPlanet, because here we use
qrot instead of mrot, but nevertheless the switch
to relying on unambiguous, physical reference
quantities greatly simplifies comparison between
different models. Setting cmp.qrot is no longer
allowed! We must set cmp.period instead. It will
be used together with cmp.mass and cmp.radius.
Note that these are all dimensioned physical
quantities, so pay attention to units. (Recommend
using MKS for everything.)

Since this is a big change, it led to reviewing
almost every file in the repo, and linting,
renaming, obsoleteing, or graveyarding many items.

* Remove dependence on extern/physunits; we don't
  need it for debugging now and we can set big G
  manually instead.

* Add some +generators; graveyard obsolete ones.

* Relocate some tests from examples/ to tests/.

* Put a placeholder for spline option in P_mid().

* Move NMoI() from access block to public block.

* Fully deprecating setting of qrot; must set
  rotation PERIOD instead!

* Lint and standardize relax_to_barotrope