Squashed commit of several tests.
As usual, it's not looking good for CMS re precision. Convergence
is never reached, with or without skip-n-spline, for as long as I
can run. Not relaxing to HE on a static profile and certainly not
relaxing to a n=1 polytrope. The good news for a barotrope run is
that we can use a small nx; the bad news is this is because the
pressure integral probably dominates as error source. As for
regular profile in HE; it's no better than tof7, and hardly better
than tof4.

The only thing that might save CMS as a practical tool is finding
a miracle layer distribution and/or some special treatment of the
upper 1 bar of atmosphere. The procedure in the Militzer paper
didn't work for me. There must be more to it than they expose in
that paper.